Bind DevilsPlayground object to reset call
The lexical context (value of this) was incorrect when the reset()
method was called. Inside reset() we were seeing the default lexical
context (the window object) instead of our DevilsPlayground instance.

By binding our reset() method to our DevilsPlayground instance (`this`
inside our addListeners() method) we correct this problem.

See #1